[InstCombine] Add !noundef if is guaranteed do not violate !range
Since D141386, violating !range would return poison. But folding select
to and/or i1 isn't poison safe in general. If we can prove that the
!range can never be violated, adding a !noundef can assume it is
guaranteed not to be posion value.

Even though violating !noundef would cause immediate undefine behavior,
in other words this instruction would not be speculatable, but metadatas
that may raise IUB is not considered in `isSafeToSpeculativelyExecute`
since these metadatas can be dropped.

Therefore, adding !noundef can improve folding of select and there is no
foreseeable side-effect at the same time.

Differential Revision: https://reviews.llvm.org/D149077

[IR] Make hasPoisonGeneratingMetadata return false if has !noundef

`hasPoisonGeneratingMetadata` is used by ValueTracking function
`canCreateUndefOrPoison`, if both poison generating metadata and
!noundef exists, then this instruction can not create undef or poison
actually.

But I am not sure if it is correct since !noundef can be dropped. If we
first do the transform based on !noundef, then we drop it in other
transforms, does it still correct?

Differential Revision: https://reviews.llvm.org/D149076